+++ java-jni/j_neo_cs.c
@@ -30,8 +30,8 @@ static void jErr(JNIEnv *env, char *error_string) {
 
 int jNeoErr (JNIEnv *env, NEOERR *err);
 
-JNIEXPORT jint JNICALL Java_org_clearsilver_CS__1init
- (JNIEnv *env, jobject obj, jint hdf_obj_ptr) {
+JNIEXPORT jlong JNICALL Java_org_clearsilver_CS__1init
+ (JNIEnv *env, jobject obj, jlong hdf_obj_ptr) {
   HDF *hdf = (HDF *)hdf_obj_ptr;
   CSPARSE *cs = NULL;
   NEOERR *err;
@@ -46,18 +46,18 @@ JNIEXPORT jint JNICALL Java_org_clearsilver_CS__1init
   err = cgi_register_strfuncs(cs);
   if (err != STATUS_OK) return jNeoErr(env,err);
 
-  return (jint) cs;
+  return (jlong) cs;
 }
 
 JNIEXPORT void JNICALL Java_org_clearsilver_CS__1dealloc
-(JNIEnv *env, jclass objClass, jint cs_obj_ptr) {
+(JNIEnv *env, jclass objClass, jlong cs_obj_ptr) {
   CSPARSE *cs = (CSPARSE *)cs_obj_ptr;
   cs_destroy(&cs);
 }
 
 
 JNIEXPORT void JNICALL Java_org_clearsilver_CS__1parseFile(JNIEnv *env,
-    jobject objCS, jint cs_obj_ptr, jstring j_filename, jboolean use_cb) {
+    jobject objCS, jlong cs_obj_ptr, jstring j_filename, jboolean use_cb) {
   CSPARSE *cs = (CSPARSE *)cs_obj_ptr;
   NEOERR *err;
   const char *filename;
@@ -90,7 +90,7 @@ JNIEXPORT void JNICALL Java_org_clearsilver_CS__1parseFile(JNIEnv *env,
 }
 
 JNIEXPORT void JNICALL Java_org_clearsilver_CS__1parseStr
-(JNIEnv *env, jclass objClass, jint cs_obj_ptr,
+(JNIEnv *env, jclass objClass, jlong cs_obj_ptr,
  jstring j_contentstring) {
 
   CSPARSE *cs = (CSPARSE *)cs_obj_ptr;
@@ -123,7 +123,7 @@ static NEOERR *render_cb (void *ctx, char *buf)
 
 
 JNIEXPORT jstring JNICALL Java_org_clearsilver_CS__1render
-(JNIEnv *env, jclass objClass, jint cs_obj_ptr) {
+(JNIEnv *env, jclass objClass, jlong cs_obj_ptr) {
   CSPARSE *cs = (CSPARSE *)cs_obj_ptr;
   STRING str;
   NEOERR *err;
@@ -173,7 +173,7 @@ JNIEXPORT jstring JNICALL Java_org_clearsilver_CS__1render
 
 // Change global HDF
 JNIEXPORT void JNICALL Java_org_clearsilver_CS__1setGlobalHdf
-(JNIEnv *env, jobject objclass, jint cs_obj_ptr, jint hdf_obj_ptr) {
+(JNIEnv *env, jobject objclass, jlong cs_obj_ptr, jlong hdf_obj_ptr) {
   HDF *hdf = (HDF *)hdf_obj_ptr;
   CSPARSE *cs = (CSPARSE *)cs_obj_ptr;
   cs->global_hdf = hdf;
